requirements.txt
click==8.1.3
Flask==2.3.0
gunicorn==20.0.4
itsdangerous
Jinja2==3.1.2
python-dotenv==1.0.0
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.0
Flask_Limiter==3.4.0
Flask_Mail==0.9.1
flask_session==0.5.0
flask_sqlalchemy==3.0.5
openai==0.27.8
PyPDF2==3.0.1
python_docx==0.8.11
redis==4.6.0
SQLAlchemy==2.0.18
Werkzeug==2.3.6
WTForms==3.0.1


